Link J3DDrawBuffer.cpp

+++ include/JSystem/J3D/J3DMaterial.h
@@ -123,7 +123,7 @@ struct J3DMaterial {
 		}
 	}
 	J3DNBTScale* getNBTScale() const { return mTexGenBlock->getNBTScale(); }
-	u32 getTexNo(u32 idx) const { return mTevBlock->getTexNo(idx); }
+	u16 getTexNo(u32 idx) const { return mTevBlock->getTexNo(idx); }
 
 	void setTevColor(u32 i, const J3DGXColorS10* i_color) { mTevBlock->setTevColor(i, i_color); }
 	void setTevKColor(u32 i, const J3DGXColor* i_color) { mTevBlock->setTevKColor(i, i_color); }
